amicci_app/serializers.py
from rest_framework import serializers
from amicci_app.models import Category


class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = ['name', 'description']
