pyproject.toml
[tool.poetry]
name = "teste-backend-amicci"
version = "0.1.0"
description = ""
authors = ["lucashamaguchi <lucas@hamaguchi.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
django-environ = "^0.11.2"
psycopg2 = "^2.9.9"
djangorestframework = "^3.14.0"
markdown = "^3.5.2"
django-filter = "^23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
